[refactor/#269-product] v1 폴더로 이동
