+++ product-service/src/main/java/com/business/productservice/application/dto/v1/request/ReqStockDecreasePostDTOApiV1.java
@@ -1,4 +1,4 @@
-package com.business.productservice.application.dto.request;
+package com.business.productservice.application.dto.v1.request;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ReqStockRestorePostDTOApiV1 {
+public class ReqStockDecreasePostDTOApiV1 {
 
     @Valid
     @NotNull(message = "재고 정보를 입력해주세요")
@@ -23,7 +23,7 @@ public class ReqStockRestorePostDTOApiV1 {
     @AllArgsConstructor
     public static class Stock {
 
-        @NotNull(message = "상품 재고 복구 수량을 입력해주세요.")
-        private Integer stockRestoreAmount;
+        @NotNull(message = "상품 재고 차감 수량을 입력해주세요.")
+        private Integer stockDecreaseAmount;
     }
 }
